backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.7.1
pymongo==4.7.2
